ReRefactor Uxd to Mailslots

+++ raet/road/test/test_keeping.py
@@ -25,6 +25,8 @@ console = getConsole()
 from raet import raeting, nacling
 from raet.road import estating, keeping, stacking
 
+def tempbasedir(prefix='', suffix='', dir='', lane='', keep=''):
+    return tempfile.mkdtemp(prefix=prefix, suffix=suffix)
 
 def setUpModule():
     console.reinit(verbosity=console.Wordage.concise)
@@ -39,11 +41,14 @@ class BasicTestCase(unittest.TestCase):
         self.store = storing.Store(stamp=0.0)
         self.timer = StoreTimer(store=self.store, duration=1.0)
 
-        self.base = tempfile.mkdtemp(prefix="raet",  suffix="base", dir='/tmp')
+        # self.base = tempfile.mkdtemp(prefix="raet",  suffix="base")
+        self.base = tempbasedir(prefix="raet", suffix='base')
 
     def tearDown(self):
-        if os.path.exists(self.base):
-            shutil.rmtree(self.base)
+
+        if not sys.platform == 'win32':
+            if os.path.exists(self.base):
+                shutil.rmtree(self.base)
 
     def createRoadData(self, name, base, auto=None):
         '''
@@ -156,9 +161,9 @@ class BasicTestCase(unittest.TestCase):
         #default ha is ("", raeting.RAET_PORT)
 
         #console.terse("{0} keep dirpath = {1}\n".format(stack.name, stack.keep.dirpath))
-        self.assertTrue(stack.keep.dirpath.endswith('/road/keep/main'))
-        self.assertTrue(stack.keep.localfilepath.endswith('/road/keep/main/local/estate.json'))
-        self.assertTrue(stack.keep.localrolepath.endswith('/road/keep/main/local/role.json'))
+        self.assertTrue(stack.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
+        self.assertTrue(stack.keep.localfilepath.endswith(os.path.join('road', 'keep', 'main', 'local', 'estate.json')))
+        self.assertTrue(stack.keep.localrolepath.endswith(os.path.join('road', 'keep', 'main', 'local', 'role.json')))
         self.assertTrue(stack.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         # test round trip
@@ -369,9 +374,9 @@ class BasicTestCase(unittest.TestCase):
         #default ha is ("", raeting.RAET_PORT)
 
         #console.terse("{0} keep dirpath = {1}\n".format(stack.name, stack.keep.dirpath))
-        self.assertTrue(stack.keep.dirpath.endswith('/road/keep/main'))
-        self.assertTrue(stack.keep.localfilepath.endswith('/road/keep/main/local/estate.msgpack'))
-        self.assertTrue(stack.keep.localrolepath.endswith('/road/keep/main/local/role.msgpack'))
+        self.assertTrue(stack.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
+        self.assertTrue(stack.keep.localfilepath.endswith(os.path.join('road', 'keep', 'main', 'local', 'estate.msgpack')))
+        self.assertTrue(stack.keep.localrolepath.endswith(os.path.join('road', 'keep', 'main', 'local', 'role.msgpack')))
         self.assertTrue(stack.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         # test round trip
@@ -573,7 +578,10 @@ class BasicTestCase(unittest.TestCase):
         fallback to ~user/.raet
         '''
         console.terse("{0}\n".format(self.testAltDirpath.__doc__))
-        base = '/var/cache/'
+        if sys.platform == 'win32':
+            base = 'c:\\var\\cache'
+        else:
+            base = '/var/cache'
         auto = raeting.autoModes.once
         data = self.createRoadData(name='main', base=base)
         keeping.clearAllKeep(data['dirpath'])
@@ -585,7 +593,7 @@ class BasicTestCase(unittest.TestCase):
         #default ha is ("", raeting.RAET_PORT)
 
         #console.terse("{0} keep dirpath = {1}\n".format(stack.name, stack.keep.dirpath))
-        self.assertTrue(".raet/keep/main" in stack.keep.dirpath)
+        self.assertTrue(os.path.join('keep', 'main') in stack.keep.dirpath)
         self.assertEqual(stack.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         # test can write
@@ -614,7 +622,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -625,7 +633,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         self.assertFalse(main.keep.auto)
 
@@ -686,7 +694,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -698,7 +706,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         self.assertFalse(main.keep.auto)
 
@@ -865,7 +873,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
         self.assertEqual(main.name, 'main')
         self.assertEqual(main.local.name, main.name)
@@ -879,7 +887,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         self.assertEqual(other.name, 'other')
         self.assertEqual(other.local.name, other.name)
@@ -983,7 +991,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
         self.assertEqual(main.name, 'main')
         self.assertEqual(main.local.name, main.name)
@@ -997,7 +1005,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         self.assertEqual(other.name, 'other')
         self.assertEqual(other.local.name, other.name)
@@ -1074,7 +1082,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -1085,7 +1093,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         self.assertIs(main.keep.auto, raeting.autoModes.once)
 
@@ -1118,7 +1126,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         # attempt to join to main with main auto accept enabled
@@ -1153,7 +1161,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         # attempt to join to main with main auto accept disabled
@@ -1211,7 +1219,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -1223,7 +1231,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         self.assertIs(main.keep.auto, raeting.autoModes.once)
 
@@ -1255,7 +1263,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         # attempt to join to main with main auto accept disabled
@@ -1313,7 +1321,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         # attempt to join to main with main auto accept disabled
@@ -1365,7 +1373,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
 
@@ -1398,7 +1406,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         # attempt to join to main with main auto accept disabled
@@ -1439,7 +1447,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=auto,
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -1450,7 +1458,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         self.join(other, main)
@@ -1483,7 +1491,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         # attempt to join to main with main auto accept enabled
@@ -1534,7 +1542,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         # attempt to join to main with main auto accept enabled and immutable
@@ -1606,7 +1614,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -1617,7 +1625,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         self.join(other, main)
@@ -1650,7 +1658,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
         remote = main.remotes.values()[0]
         self.assertEqual(remote.acceptance, raeting.acceptances.accepted) # saved still accepted
@@ -1700,7 +1708,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         # attempt to join to main with main auto accept enabled
@@ -1762,7 +1770,7 @@ class BasicTestCase(unittest.TestCase):
                                      ha=None)
         #default ha is ("", raeting.RAET_PORT)
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
 
         data = self.createRoadData(name='other', base=self.base)
@@ -1774,7 +1782,7 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
 
         self.join(other, main)
@@ -1819,12 +1827,12 @@ class BasicTestCase(unittest.TestCase):
                                      auto=None,
                                      ha=("", raeting.RAET_TEST_PORT))
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
         remote = main.remotes.values()[0]
         self.assertEqual(remote.acceptance, raeting.acceptances.accepted) # saved still accepted
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         remote = other.remotes.values()[0]
         self.assertEqual(remote.acceptance, raeting.acceptances.accepted) # saved still accepted
@@ -1887,12 +1895,12 @@ class BasicTestCase(unittest.TestCase):
             #other.addRemote(remote)
         #other.dumpRemotes()
 
-        self.assertTrue(main.keep.dirpath.endswith('road/keep/main'))
+        self.assertTrue(main.keep.dirpath.endswith(os.path.join('road', 'keep', 'main')))
         self.assertEqual(main.local.ha, ("0.0.0.0", raeting.RAET_PORT))
         remote = main.remotes.values()[0]
         self.assertEqual(remote.acceptance, raeting.acceptances.accepted) # saved still accepted
 
-        self.assertTrue(other.keep.dirpath.endswith('road/keep/other'))
+        self.assertTrue(other.keep.dirpath.endswith(os.path.join('road', 'keep', 'other')))
         self.assertEqual(other.local.ha, ("0.0.0.0", raeting.RAET_TEST_PORT))
         remote = other.remotes.values()[0]
         self.assertEqual(remote.acceptance, raeting.acceptances.accepted) # saved still accepted
